introduce TesseractPath
