+++ examples/pathreference/tesseract_api.py
@@ -9,6 +9,7 @@ from typing import Annotated
 from pydantic import AfterValidator, BaseModel
 
 from tesseract_core.runtime.config import get_config
+from tesseract_core.runtime.experimental import TesseractPath
 
 
 def bin_reference(path: Path) -> str | None:
@@ -37,11 +38,11 @@ def has_bin_sidecar(path: Path) -> Path:
 
 
 class InputSchema(BaseModel):
-    paths: list[Annotated[Path, AfterValidator(has_bin_sidecar)]]
+    paths: list[Annotated[TesseractPath, AfterValidator(has_bin_sidecar)]]
 
 
 class OutputSchema(BaseModel):
-    paths: list[Annotated[Path, AfterValidator(has_bin_sidecar)]]
+    paths: list[Annotated[TesseractPath, AfterValidator(has_bin_sidecar)]]
 
 
 def apply(inputs: InputSchema) -> OutputSchema:
