+++ tesseract_core/runtime/file_interactions.py
@@ -122,57 +122,3 @@ def parent_path(x: PathLike) -> PathLike:
         return urllib.parse.urljoin(x, "..")
 
     return type(x)(Path(x).parent.as_posix())
-
-
-def resolve_input_path(path: Path) -> Path:
-    """Resolve a relative path against RuntimeConfig().input_path.
-
-    Validates that the resolved path stays within the input directory and exists.
-
-    Raises:
-        ValueError: If the path escapes the input directory.
-        FileNotFoundError: If the resolved path does not exist.
-    """
-    from tesseract_core.runtime.config import get_config
-
-    input_path = get_config().input_path
-    tess_path = (input_path / path).resolve()
-    if str(input_path) not in str(tess_path):
-        raise ValueError(
-            f"Invalid input file reference: {path}. "
-            f"Expected path to be relative to {input_path}, but got {tess_path}. "
-            "File references have to be relative to --input-path."
-        )
-    if not tess_path.exists():
-        raise FileNotFoundError(f"Input path {tess_path} does not exist.")
-    return tess_path
-
-
-def strip_output_path(path: Path) -> Path:
-    """Strip RuntimeConfig().output_path prefix from a path.
-
-    If the path is relative to the output directory, returns the relative portion.
-
-    Raises:
-        ValueError: If the path does not exist or is outside the output directory.
-    """
-    from tesseract_core.runtime.config import get_config
-
-    output_path = get_config().output_path
-    if path.is_relative_to(output_path):
-        if not path.exists():
-            raise ValueError(f"Output path {path} does not exist inside Tesseract")
-        return path.relative_to(output_path)
-    else:
-        full_path = output_path / path
-        if not full_path.exists():
-            if path.exists():
-                raise ValueError(
-                    f"Output path {path} is not in {output_path}. "
-                    f"All output data must be copied to `--output-path` ({output_path})."
-                )
-            else:
-                raise ValueError(
-                    f"Output path {path} is not in {output_path} or Tesseract root"
-                )
-        return path
